+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [#116] continue installing components when the component operator restarts while installing/upgrading
 - [#119] add german translation for proxy_configuration.md
 
 ## [v1.14.0] - 2026-05-20